Conflicts:
	remote-tenant-resolver/src/main/java/fr/cnes/regards/microservices/administration/MicroserviceTenantConnectionResolver.java

+++ rs-admin/accessrights/accessrights-service/src/main/java/fr/cnes/regards/modules/accessrights/service/RoleService.java
@@ -15,23 +15,29 @@ import javax.annotation.Resource;
 
 import org.springframework.stereotype.Service;
 
+import fr.cnes.regards.framework.module.rest.exception.AlreadyExistingException;
+import fr.cnes.regards.framework.module.rest.exception.EntityNotFoundException;
+import fr.cnes.regards.framework.module.rest.exception.InvalidValueException;
+import fr.cnes.regards.framework.module.rest.exception.OperationForbiddenException;
 import fr.cnes.regards.modules.accessrights.dao.projects.IRoleRepository;
 import fr.cnes.regards.modules.accessrights.domain.projects.ProjectUser;
 import fr.cnes.regards.modules.accessrights.domain.projects.ResourcesAccess;
 import fr.cnes.regards.modules.accessrights.domain.projects.Role;
-import fr.cnes.regards.modules.core.exception.AlreadyExistingException;
-import fr.cnes.regards.modules.core.exception.EntityNotFoundException;
-import fr.cnes.regards.modules.core.exception.InvalidValueException;
 import fr.cnes.regards.modules.core.utils.RegardsStreamUtils;
 
 /**
  * {@link IRoleService} implementation
  *
- * @author xbrochar
+ * @author Xavier-Alexandre Brochard
  *
  */
 @Service
 public class RoleService implements IRoleService {
+
+    /**
+     * Error message
+     */
+    private static final String NATIVE_ROLE_NOT_REMOVABLE = "Modifications on native roles are forbidden";
 
     /**
      * CRUD repository managing {@link Role}s. Autowired by Spring.
@@ -88,7 +94,7 @@ public class RoleService implements IRoleService {
 
     @Override
     public void updateRole(final Long pRoleId, final Role pUpdatedRole)
-            throws InvalidValueException, EntityNotFoundException {
+            throws EntityNotFoundException, InvalidValueException {
         if (!pRoleId.equals(pUpdatedRole.getId())) {
             throw new InvalidValueException();
         }
@@ -99,7 +105,11 @@ public class RoleService implements IRoleService {
     }
 
     @Override
-    public void removeRole(final Long pRoleId) {
+    public void removeRole(final Long pRoleId) throws OperationForbiddenException {
+        final Role previous = roleRepository.findOne(pRoleId);
+        if ((previous != null) && previous.isNative()) {
+            throw new OperationForbiddenException(pRoleId.toString(), Role.class, NATIVE_ROLE_NOT_REMOVABLE);
+        }
         roleRepository.delete(pRoleId);
     }
